not working/gmgn_wallet_filter.py
import asyncio
import json
import os
import logging
import argparse
import re
from pathlib import Path
import time
from datetime import datetime
from typing import Dict, List, Set, Tuple
import sys
from bs4 import BeautifulSoup
import traceback

# Add project root to Python path
project_root = str(Path(__file__).parent.parent)
if project_root not in sys.path:
    sys.path.append(project_root)

# Import the GMGN scraper
from scripts.gmgn_hybrid import CloudflareBypass

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

class GMGNWalletFilter:
    def __init__(self, input_file: str = None, output_file: str = None):
        self.input_file = input_file or os.path.join(project_root, 'scripts', 'wallet_addresses.txt')
        self.output_file = output_file or 'data/filtered_wallets.json'
    
    async def filter_wallets(self, target_count: int = 10) -> List[Dict]:
        """
        Filter wallets based on GMGN data
        
        Args:
            target_count (int): Number of suitable wallets to find
            
        Returns:
            List[Dict]: List of suitable wallets with their data
        """
        logger.info(f"Starting wallet filtering process... Target count: {target_count}")
        filtered_wallets = []
        
        # Read wallet addresses
        try:
            with open(self.input_file, 'r') as f:
                wallet_addresses = [line.strip() for line in f if line.strip()]
        except Exception as e:
            logger.error(f"Error reading wallet addresses: {e}")
            return []
        
        logger.info(f"Found {len(wallet_addresses)} wallet addresses to process")
        
        # Create scraper instance
        scraper = CloudflareBypass(headless=True)
        
        for wallet_address in wallet_addresses:
            try:
                # Process wallet
                result = await scraper.process_wallet(
                    wallet_address=wallet_address,
                    cloudflare_wait=30
                )
                
                if result['success']:
                    filtered_wallets.append(result['data'])
                    logger.info(f"Successfully processed wallet {wallet_address}")
                    
                    # Check if we've found enough wallets
                    if len(filtered_wallets) >= target_count:
                        logger.info(f"Found {target_count} suitable wallets, stopping search")
                        break
                else:
                    logger.warning(f"Failed to process wallet {wallet_address}: {result.get('error')}")
                    
            except Exception as e:
                logger.error(f"Error processing wallet: {str(e)}")
                continue
            
            # Add delay between requests
            await asyncio.sleep(5)
        
        # Save results
        if filtered_wallets:
            try:
                os.makedirs(os.path.dirname(self.output_file), exist_ok=True)
                with open(self.output_file, 'w') as f:
                    json.dump(filtered_wallets, f, indent=2)
                logger.info(f"Saved {len(filtered_wallets)} filtered wallets to {self.output_file}")
            except Exception as e:
                logger.error(f"Error saving filtered wallets: {e}")
        
        return filtered_wallets

async def main():
    # Parse command line arguments
    parser = argparse.ArgumentParser(description='Filter wallets using GMGN data')
    parser.add_argument('--input', type=str, help='Input file with wallet addresses')
    parser.add_argument('--output', type=str, help='Output file for filtered wallets')
    parser.add_argument('--target-count', type=int, default=10,
                        help='Number of suitable wallets to find')
    args = parser.parse_args()
    
    # Create filter instance
    wallet_filter = GMGNWalletFilter(
        input_file=args.input,
        output_file=args.output
    )
    
    # Filter wallets
    suitable_wallets = await wallet_filter.filter_wallets(args.target_count)
    
    print(f"\nFiltering complete!")
    print(f"Found {len(suitable_wallets)} suitable wallets")
    if args.output:
        print(f"Results saved to {args.output}")

if __name__ == "__main__":
    try:
        asyncio.run(main())
    except KeyboardInterrupt:
        print("\nShutting down...")
    except Exception as e:
        print(f"Error: {str(e)}")
        print(f"Traceback: {traceback.format_exc()}")