requirements.txt
aiohttp==3.11.11
discord.py==2.4.0
pandas==2.2.3
python-dotenv==1.1.0
pytz==2024.2
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scripts==3.0
Telethon==1.38.1
